[BE] REST Docs로 API 명세 작성 (#242)
* build: build.gradle에 Rest DOCS 의존성 추가 #241

- 테스트 실행 전 snippet을 다 삭제하는 태스크를 추가했습니다.

* feat: RestDocs 설정 #241

- default snippets 중에서 `curl-request`와 `httpie-request`를 사용하지 않아 수정했습니다.
- request, response에 pretty print 옵션을 넣어서 json 형식을 출력합니다.
- util 패키지에 IntegrationTest를 넣었습니다.

* docs: request-fields, response-fields 템플릿 생성 #241

- `optional()`을 표시하기 위해 템플릿을 만들었습니다.

* feat: 테스트 코드에 Rest Docs 적용 #241

- 통합테스트를 문서화를 담당하는 패키지로 `documentation`, 그 외 여러 인수 테스트를 담당하는 `scenario` 패키지로 구분했습니다.
- 유닛 테스트 패키지도 만들었습니다.

* docs: index.adoc 작성 #241

- 스니펫을 모아서 최종 문서화를 했습니다.

* chore: JWT secret key 변경

- 보안을 위해 업데이트 했습니다.
