+++ be/src/test/java/codesquad/bookkbookk/BookkbookkApplicationTests.java
@@ -2,7 +2,9 @@ package codesquad.bookkbookk;
 
 import org.junit.jupiter.api.Test;
 
-class BookkbookkApplicationTests extends IntegrationTest{
+import codesquad.bookkbookk.util.IntegrationTest;
+
+class BookkbookkApplicationTests extends IntegrationTest {
 
     @Test
     void contextLoads() {
