+++ be/src/test/java/codesquad/bookkbookk/integration/scenario/AuthTest.java
@@ -1,0 +1,64 @@
+package codesquad.bookkbookk.integration.scenario;
+
+import org.assertj.core.api.SoftAssertions;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.ResponseCookie;
+
+import codesquad.bookkbookk.common.error.exception.RefreshTokenNotSavedException;
+import codesquad.bookkbookk.common.jwt.JwtProperties;
+import codesquad.bookkbookk.common.jwt.JwtProvider;
+import codesquad.bookkbookk.domain.member.data.entity.Member;
+import codesquad.bookkbookk.domain.member.repository.MemberRepository;
+import codesquad.bookkbookk.util.IntegrationTest;
+import codesquad.bookkbookk.util.TestDataFactory;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+
+public class AuthTest extends IntegrationTest {
+
+    @Autowired
+    private MemberRepository memberRepository;
+    @Autowired
+    private JwtProvider jwtProvider;
+    @Autowired
+    private JwtProperties jwtProperties;
+
+    @Test
+    @DisplayName("refresh token이 redis에 없으면 accessToken 재발급이 실패한다.")
+    void reissueAccessTokenWithExpiredRefreshToken() throws InterruptedException {
+        // given
+        Member member = TestDataFactory.createMember();
+        memberRepository.save(member);
+
+        String refreshToken = jwtProvider.createRefreshToken("uuid");
+
+        ResponseCookie cookie = ResponseCookie.from("refreshToken", refreshToken)
+                .domain("localhost")
+                .path("/")
+                .maxAge(jwtProperties.getRefreshTokenExpiration() / 1000)
+                .build();
+
+        RefreshTokenNotSavedException exception = new RefreshTokenNotSavedException();
+
+        // when
+        ExtractableResponse<Response> response = RestAssured
+                .given().log().all()
+                .header(HttpHeaders.COOKIE, cookie.toString())
+                .when()
+                .post("/api/auth/reissue")
+                .then().log().all()
+                .extract();
+
+        // then
+        SoftAssertions.assertSoftly(assertions -> {
+            assertions.assertThat(response.statusCode()).isEqualTo(exception.getStatus().value());
+            assertions.assertThat(response.jsonPath().getString("message")).isEqualTo(exception.getMessage());
+        });
+    }
+
+}
