+++ be/src/test/java/codesquad/bookkbookk/integration/scenario/MemberTest.java
@@ -1,0 +1,47 @@
+package codesquad.bookkbookk.integration.scenario;
+
+import org.assertj.core.api.SoftAssertions;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+
+import codesquad.bookkbookk.common.error.exception.ApiException;
+import codesquad.bookkbookk.common.error.exception.MemberNotFoundException;
+import codesquad.bookkbookk.common.jwt.JwtProvider;
+import codesquad.bookkbookk.util.IntegrationTest;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+
+public class MemberTest extends IntegrationTest {
+
+    @Autowired
+    private JwtProvider jwtProvider;
+
+    @Test
+    @DisplayName("저장되어 있지 않은 멤버를 조회하면 예외를 던진다.")
+    void readUnsavedMember() {
+        //given
+        String accessToken = jwtProvider.createAccessToken(1L);
+        MemberNotFoundException exception = new MemberNotFoundException();
+
+        //when
+        ExtractableResponse<Response> response = RestAssured
+                .given().log().all()
+                .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
+                .when()
+                .get("/api/members")
+                .then().log().all()
+                .extract();
+
+        //then
+        SoftAssertions.assertSoftly(softAssertions -> {
+            softAssertions.assertThat(response.statusCode()).isEqualTo(exception.getStatus().value());
+            softAssertions.assertThat(response.jsonPath().getObject("", ApiException.class).getMessage())
+                    .isEqualTo(exception.getMessage());
+        });
+    }
+
+}
