--- be/src/test/java/codesquad/bookkbookk/topic/integration/TopicTest.java
@@ -1,235 +1,0 @@
-package codesquad.bookkbookk.topic.integration;
-
-import java.util.List;
-
-import org.assertj.core.api.SoftAssertions;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-
-import codesquad.bookkbookk.IntegrationTest;
-import codesquad.bookkbookk.common.error.exception.TopicNotFoundException;
-import codesquad.bookkbookk.common.jwt.JwtProvider;
-import codesquad.bookkbookk.domain.book.data.entity.Book;
-import codesquad.bookkbookk.domain.book.repository.BookRepository;
-import codesquad.bookkbookk.domain.bookclub.data.entity.BookClub;
-import codesquad.bookkbookk.domain.bookclub.repository.BookClubRepository;
-import codesquad.bookkbookk.domain.bookmark.data.entity.Bookmark;
-import codesquad.bookkbookk.domain.bookmark.repository.BookmarkRepository;
-import codesquad.bookkbookk.domain.chapter.data.entity.Chapter;
-import codesquad.bookkbookk.domain.chapter.repository.ChapterRepository;
-import codesquad.bookkbookk.domain.mapping.entity.BookClubMember;
-import codesquad.bookkbookk.domain.mapping.repository.BookClubMemberRepository;
-import codesquad.bookkbookk.domain.member.data.entity.Member;
-import codesquad.bookkbookk.domain.member.repository.MemberRepository;
-import codesquad.bookkbookk.domain.topic.data.dto.UpdateTopicTitleRequest;
-import codesquad.bookkbookk.domain.topic.data.entity.Topic;
-import codesquad.bookkbookk.domain.topic.repository.TopicRepository;
-import codesquad.bookkbookk.util.TestDataFactory;
-
-import io.restassured.RestAssured;
-import io.restassured.http.ContentType;
-import io.restassured.response.ExtractableResponse;
-import io.restassured.response.Response;
-
-public class TopicTest extends IntegrationTest {
-
-    @Autowired
-    private MemberRepository memberRepository;
-
-    @Autowired
-    private TopicRepository topicRepository;
-
-    @Autowired
-    private ChapterRepository chapterRepository;
-
-    @Autowired
-    private BookRepository bookRepository;
-
-    @Autowired
-    private BookClubRepository bookClubRepository;
-
-    @Autowired
-    private BookClubMemberRepository bookClubMemberRepository;
-
-    @Autowired
-    private BookmarkRepository bookmarkRepository;
-
-    @Autowired
-    private JwtProvider jwtProvider;
-
-    @DisplayName("성공적으로 토픽을 생성한다.")
-    @Test
-    void createTopic(){
-        //given
-        Member member = TestDataFactory.createMember();
-        memberRepository.save(member);
-        String accessToken = jwtProvider.createAccessToken(member.getId());
-
-        BookClub bookClub = TestDataFactory.createBookClub(member);
-        bookClubRepository.save(bookClub);
-
-        BookClubMember bookClubMember = new BookClubMember(bookClub, member);
-        bookClubMemberRepository.save(bookClubMember);
-
-        Book book = TestDataFactory.createBook(bookClub);
-        bookRepository.save(book);
-
-        Chapter chapter = new Chapter(book, "first");
-        chapterRepository.save(chapter);
-
-        String requestBody = "{\"chapterId\": 1, \"title\": \"토픽\"}";
-        //when
-        ExtractableResponse<Response> response = RestAssured
-                .given().log().all()
-                .body(requestBody)
-                    .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
-                    .contentType(ContentType.JSON)
-                .when()
-                    .post("/api/topics")
-                    .then().log().all()
-                .extract();
-
-        //then
-        SoftAssertions.assertSoftly(softAssertions -> {
-            softAssertions.assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-            softAssertions.assertThat(response.jsonPath().getLong("createdTopicId")).isNotZero();
-        });
-
-    }
-
-    @DisplayName("토픽의 북마크들을 조회한다.")
-    @Test
-    void readBookmarks(){
-        //given
-        Member member = TestDataFactory.createMember();
-        memberRepository.save(member);
-        String accessToken = jwtProvider.createAccessToken(member.getId());
-
-        BookClub bookClub = TestDataFactory.createBookClub(member);
-        bookClubRepository.save(bookClub);
-
-        BookClubMember bookClubMember = new BookClubMember(bookClub, member);
-        bookClubMemberRepository.save(bookClubMember);
-
-        Book book = TestDataFactory.createBook(bookClub);
-        bookRepository.save(book);
-
-        Chapter chapter = new Chapter(book, "first");
-        chapterRepository.save(chapter);
-
-        Topic topic = new Topic(chapter, "topic");
-        topicRepository.save(topic);
-
-        List<Bookmark> bookmarks = TestDataFactory.createBookmarks(5, member, topic);
-        bookmarkRepository.saveAll(bookmarks);
-
-        //when
-        ExtractableResponse<Response> response = RestAssured
-                .given().log().all()
-                .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
-                .when()
-                .get("/api/topics/" + topic.getId() + "/bookmarks")
-                .then().log().all()
-                .extract();
-
-        //then
-        SoftAssertions.assertSoftly(softAssertions -> {
-            softAssertions.assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-            softAssertions.assertThat(response.jsonPath().getList("").size()).isEqualTo(5);
-            softAssertions.assertThat(response.jsonPath().getString("[1].author.nickname"))
-                    .isEqualTo("nickname");
-            softAssertions.assertThat(response.jsonPath().getLong("[3].bookmarkId"))
-                    .isEqualTo(4);
-        });
-
-    }
-
-
-
-    @DisplayName("성공적으로 토픽의 제목을 수정한다")
-    @Test
-    void updateTitle() {
-        //given
-        Member member = TestDataFactory.createMember();
-        memberRepository.save(member);
-        String accessToken = jwtProvider.createAccessToken(member.getId());
-
-        BookClub bookClub = TestDataFactory.createBookClub(member);
-        bookClubRepository.save(bookClub);
-
-        BookClubMember bookClubMember = new BookClubMember(bookClub, member);
-        bookClubMemberRepository.save(bookClubMember);
-
-        Book book = TestDataFactory.createBook(bookClub);
-        bookRepository.save(book);
-
-        Chapter chapter = new Chapter(book, "first");
-        chapterRepository.save(chapter);
-
-        Topic topic = new Topic(chapter, "topic");
-        topicRepository.save(topic);
-
-        UpdateTopicTitleRequest request = new UpdateTopicTitleRequest("updated title");
-
-        //when
-        ExtractableResponse<Response> response = RestAssured
-                .given().log().all()
-                    .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
-                    .contentType(ContentType.JSON)
-                    .body(request)
-                .when()
-                    .patch("/api/topics/"+ topic.getId())
-                .then().log().all()
-                    .extract();
-
-        //then
-        Topic result = topicRepository.findById(topic.getId()).orElseThrow(TopicNotFoundException::new);
-        SoftAssertions.assertSoftly(softAssertions -> {
-            softAssertions.assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-            softAssertions.assertThat(result.getTitle()).isEqualTo("updated title");
-        });
-    }
-
-    @DisplayName("성공적으로 토픽을 삭제한다")
-    @Test
-    void deleteTopic() {
-        //given
-        Member member = TestDataFactory.createMember();
-        memberRepository.save(member);
-        String accessToken = jwtProvider.createAccessToken(member.getId());
-
-        BookClub bookClub = TestDataFactory.createBookClub(member);
-        bookClubRepository.save(bookClub);
-
-        BookClubMember bookClubMember = new BookClubMember(bookClub, member);
-        bookClubMemberRepository.save(bookClubMember);
-
-        Book book = TestDataFactory.createBook(bookClub);
-        bookRepository.save(book);
-
-        Chapter chapter = new Chapter(book, "first");
-        chapterRepository.save(chapter);
-
-        Topic topic = new Topic(chapter, "topic");
-        topicRepository.save(topic);
-
-        //when
-        ExtractableResponse<Response> response = RestAssured
-                .given().log().all()
-                .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
-                .contentType(ContentType.JSON)
-                .when()
-                .delete("/api/topics/"+ topic.getId())
-                .then().log().all()
-                .extract();
-
-        //then
-        SoftAssertions.assertSoftly(softAssertions -> {
-            softAssertions.assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-            softAssertions.assertThat(topicRepository.findById(topic.getId())).isEmpty();
-        });
-    }
-}
