+++ be/src/test/resources/org/springframework/restdocs/templates/request-fields.snippet
@@ -1,0 +1,11 @@
+|===
+|필드명|타입|필수값|양식|설명
+
+{{#fields}}
+    |{{#tableCellContent}}`+{{path}}+`{{/tableCellContent}}
+    |{{#tableCellContent}}`+{{type}}+`{{/tableCellContent}}
+    |{{#tableCellContent}}{{^optional}}true{{/optional}}{{/tableCellContent}}
+    |{{#tableCellContent}}{{#format}}{{.}}{{/format}}{{/tableCellContent}}
+    |{{#tableCellContent}}{{description}}{{/tableCellContent}}
+{{/fields}}
+|===
