+++ be/src/test/resources/org/springframework/restdocs/templates/response-fields.snippet
@@ -1,0 +1,10 @@
+|===
+|필드명|타입|필수값|설명
+
+{{#fields}}
+    |{{#tableCellContent}}`+{{path}}+`{{/tableCellContent}}
+    |{{#tableCellContent}}`+{{type}}+`{{/tableCellContent}}
+    |{{#tableCellContent}}{{^optional}}true{{/optional}}{{/tableCellContent}}
+    |{{#tableCellContent}}{{description}}{{/tableCellContent}}
+{{/fields}}
+|===
